requirements.txt
absl-py==2.0.0
annotated-types==0.7.0
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
branca==0.7.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
dash==2.15.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.115.6
fastapi-cli==0.0.7
fastjsonschema==2.19.0
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
folium==0.15.1
fonttools==4.46.0
fqdn==1.5.1
frozendict==2.4.0
gast==0.4.0
google-auth==2.24.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.59.3
h11==0.14.0
h5py==3.12.1
html5lib==1.1
httpcore==1.0.5
httptools==0.6.4
httpx==0.27.0
idna==3.6
imbalanced-learn==0.12.4
imblearn==0.0
importlib-metadata==7.0.1
ipykernel==6.27.1
ipython==8.18.1
ipython-genutils==0.2.0
ipython-sql==0.5.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jiter==0.8.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.11.1
jupyter_server_terminals==0.4.4
jupyterlab==4.1.6
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==3.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==16.0.6
lxml==5.1.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
mpmath==1.3.0
multitasking==0.0.11
mysql-connector-python==9.2.0
namex==0.0.8
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.4
oauthlib==3.2.2
openai==1.60.0
opencv-contrib-python==4.7.0.72
opencv-python==4.10.0.84
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.13.1
overrides==7.4.0
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.3
patsy==1.0.1
Pillow==10.1.0
platformdirs==4.0.0
plotly==5.18.0
prettytable==3.9.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
protobuf==4.25.5
psutil==5.9.6
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.10.5
pydantic_core==2.27.2
pygame==2.6.1
Pygments==2.17.2
pymongo==4.10.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.20
pytz==2023.3.post1
PyWavelets==1.8.0
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.31.1
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.13.2
rsa==4.9
scikit-learn==1.4.0
scipy==1.11.4
scratch==1.0.0
seaborn==0.13.2
Send2Trash==1.8.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.41.3
statsmodels==0.14.4
sympy==1.12
tenacity==8.2.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.18.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.4.0
terminado==0.18.0
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4
tqdm==4.67.1
traitlets==5.14.0
typer==0.15.1
types-python-dateutil==2.8.19.14
typing_extensions==4.12.2
tzdata==2023.4
uri-template==1.3.0
urllib3==2.1.0
uvicorn==0.34.0
watchfiles==1.0.4
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==14.2
Werkzeug==3.0.1
wget==3.2
widgetsnbextension==4.0.9
wrapt==1.14.1
xyzservices==2023.10.1
yfinance==0.2.4
zipp==3.17.0
